requirements.txt

flask
openai>=1.3.0
python-dotenv
openai-agents
gunicorn
flask-cors